+++ src/Pages/general/computePy.js
@@ -21,8 +21,8 @@ window.computePy_func = function computePy_func({ e = null, pyfile = "", args = 
 
         let target;
         if (!general) {
-            target = e.target
-            target.classList.toggle("is-loading")
+            target = e?.target
+            target?.classList.toggle("is-loading")
 
         }
 
@@ -43,7 +43,7 @@ window.computePy_func = function computePy_func({ e = null, pyfile = "", args = 
                     
                     
                         const pyEvent = new CustomEvent('pyEvent', { bubbles: false, detail: { py, pyfile } });
-                        e.target.dispatchEvent(pyEvent)
+                        e?.target.dispatchEvent(pyEvent)
                         console.log("pyEvent dispatched")
                     }
 
@@ -55,7 +55,7 @@ window.computePy_func = function computePy_func({ e = null, pyfile = "", args = 
 
                         if(e) {
                             const pyEventClosed = new CustomEvent('pyEventClosed', { bubbles: false, detail: { py, pyfile } });
-                            e.target.dispatchEvent(pyEventClosed)
+                            e?.target.dispatchEvent(pyEventClosed)
                             console.log("pyEventClosed dispatched")
                         }
                      })
@@ -67,7 +67,7 @@ window.computePy_func = function computePy_func({ e = null, pyfile = "", args = 
                         if(e) {
                     
                             const pyEventData = new CustomEvent('pyEventData', { bubbles: false, detail: { py, pyfile, dataReceived } });
-                            e.target.dispatchEvent(pyEventData)
+                            e?.target.dispatchEvent(pyEventData)
                          
                             console.log("pyEventData dispatched")
     
@@ -87,7 +87,7 @@ window.computePy_func = function computePy_func({ e = null, pyfile = "", args = 
                     if(e) {
                     
                         const pyEvent = new CustomEvent('pyEvent', { bubbles: false, detail: { py, pyfile } });
-                        e.target.dispatchEvent(pyEvent)
+                        e?.target.dispatchEvent(pyEvent)
                         console.log("pyEvent dispatched")
 
                     }
@@ -102,7 +102,7 @@ window.computePy_func = function computePy_func({ e = null, pyfile = "", args = 
                         if(e) {
                     
                             const pyEventData = new CustomEvent('pyEventData', { bubbles: false, detail: { py, pyfile, dataReceived } });
-                            e.target.dispatchEvent(pyEventData)
+                            e?.target.dispatchEvent(pyEventData)
                          
                             console.log("pyEventData dispatched")
     
@@ -130,7 +130,7 @@ window.computePy_func = function computePy_func({ e = null, pyfile = "", args = 
 
                         if(e) {
                             const pyEventClosed = new CustomEvent('pyEventClosed', { bubbles: false, detail: { py, pyfile } });
-                            e.target.dispatchEvent(pyEventClosed)
+                            e?.target.dispatchEvent(pyEventClosed)
                         
                             console.log("pyEventClosed dispatched")
                         }