+++ src/Pages/normline/widgets/preprocessing/TheoryRow.svelte
@@ -1,24 +1,27 @@
 
 <script>
-    import {toggleRow} from "../../functions/svelteWritables";
+
+    import {toggleRow, felixopoLocation} from "../../functions/svelteWritables";
     import Textfield from '@smui/textfield';
     import QuickBrowser from '../../../../components/QuickBrowser.svelte';
     import { fade } from 'svelte/transition';
+    
     import {theory_func} from '../../functions/theory';
 
-    export let theoryLocation, currentLocation, show_theoryplot, normMethod, preModal;
+    import CustomSwitch from '../../../../components/CustomSwitch.svelte';
+    export let theoryLocation, show_theoryplot, normMethod, preModal;
 
-    let sigma=20, scale=1, theoryfiles=[]
+    let sigma=7, scale=1, theoryfiles=[], tkplot=false;
     let showTheoryFiles = false, theoryfilesChecked = []
-
     $: if(fs.existsSync(theoryLocation)) { theoryfiles =theoryfilesChecked.map(file=>path.resolve(theoryLocation, file)) }
 
-    function plotData({e=null, tkplot="run"}={}){
+    function plotData(e=null){
         let pyfile="theory.py", args;
         
         if(theoryfiles.length < 1) return window.createToast("No files selected", "danger")
         
-        args=[...theoryfiles, normMethod, sigma, scale, currentLocation, tkplot]
+        args={theoryfiles, normMethod, sigma, scale, currentLocation:$felixopoLocation, tkplot, onlyExpRange}
+        args=[JSON.stringify(args)]
         computePy_func({e, pyfile, args})
         .then((dataFromPython)=>{
             theory_func({dataFromPython, normMethod})
@@ -27,17 +30,23 @@
         }).catch(err=>{preModal.modalContent = err;  preModal.open = true})
     }
 
+    let onlyExpRange = true;
 </script>
 
-<QuickBrowser title="Theory files" bind:active={showTheoryFiles} bind:currentLocation={theoryLocation} bind:fileChecked={theoryfilesChecked} on:submit="{(e)=>{plotData({e:e.detail.event}); localStorage["theoryLocation"] = theoryLocation}}"/>
+<QuickBrowser title="Theory files" bind:active={showTheoryFiles} bind:currentLocation={theoryLocation} bind:fileChecked={theoryfilesChecked} on:submit="{(e)=>{plotData(e.detail.event); localStorage["theoryLocation"] = theoryLocation}}"/>
 
 {#if $toggleRow}
-
     <div class="align" transition:fade>
 
         <button class="button is-link" on:click="{()=>{showTheoryFiles = !showTheoryFiles;}}"> Browse File</button>
         <Textfield type="number" style="width:7em; margin-right:0.5em;" variant="outlined" bind:value={sigma} label="Sigma"/>
+
+
         <Textfield type="number" style="width:7em" variant="outlined" bind:value={scale} label="Scale"/>
-        <button class="button is-link" on:click="{(e)=>plotData({e:e})}">Replot</button>
+        
+        <CustomSwitch style="margin: 0 1em;" bind:selected={onlyExpRange} label="Only Exp. Range"/>
+        <CustomSwitch style="margin: 0 1em;" bind:selected={tkplot} label="Matplotlib"/>
+        
+        <button class="button is-link" on:click="{plotData}">Replot</button>
     </div>
 {/if}