+++ src/components/Layout.svelte
@@ -14,7 +14,11 @@
                 properties: [type, "multiSelections"],
             }
 
-            if (process.versions.electron >= "7") {
+
+            const version = parseInt(process.versions.electron.split(".")[0])
+
+            
+            if (version >= 7) {
                 remote.dialog.showOpenDialog(mainWindow, options)
                 .then(result => {
                     console.log(result.canceled)
@@ -55,13 +59,22 @@
 
     ////////////////////////////////////////////////////////////////////////////
 
-    export let id, fileChecked=[], filetype = "felix", toggleBrowser = false, preModal = {};
+    export let id, fileChecked=[], filetype = "felix", toggleBrowser = false, preModal = {}, fullfileslist = [];
     export let currentLocation = localStorage[`${filetype}_location`] || "";
     const dispatch = createEventDispatcher()
 
     function browse_folder() {
         browse({dir:true}).then(result=>{
-            if (!result.canceled) { currentLocation= localStorage[`${filetype}_location`] = result.filePaths[0] }
+
+            console.log(result, currentLocation)
+
+
+            if (!result.canceled) { 
+                currentLocation= localStorage[`${filetype}_location`] = result.filePaths[0]
+
+                console.log(result, currentLocation)
+
+             }
         })
     }
 
@@ -77,6 +90,7 @@
         plotContainer.style.height = `calc(${ContainerHeight}px - ${buttonContainerHeight}px - 11em)`
     
     });
+
 </script>
 
 <style lang="scss">
@@ -162,12 +176,11 @@
 
 
 <section {id} style="display:none" class="animated fadeIn">
-
     <div class="columns">
 
         {#if toggleBrowser}
             <div class="column is-one-fifth-widescreen is-one-quarter-desktop box filebrowser adjust-right" transition:fly="{{ x: -100, duration: 500 }}">
-                <FileBrowser bind:currentLocation {filetype} bind:fileChecked on:chdir/>
+                <FileBrowser bind:currentLocation {filetype} bind:fileChecked on:chdir bind:fullfileslist/>
             </div>
         {/if}